Modifying allocation domain and sql extension implementation
- generalizing code for Scan and Value functions for sql types to persist PostgreSQL JSONB type
- refactoring and cleaning code for domain layer allocation aggregate
