+++ src/main/go/domain/infra/repository/allocation_plan_repository_model.go
@@ -3,7 +3,7 @@ package repository
 import (
 	"github.com/benizzio/open-asset-allocator/domain"
 	"github.com/benizzio/open-asset-allocator/domain/allocation"
-	"github.com/benizzio/open-asset-allocator/infra/sqlext"
+	"github.com/benizzio/open-asset-allocator/infra/rdbms/sqlext"
 	"github.com/benizzio/open-asset-allocator/langext"
 	"github.com/shopspring/decimal"
 )
