+++ src/main/go/infra/rdbms/sqlext/sqlext_null_time.go
@@ -1,0 +1,15 @@
+package sqlext
+
+import (
+	"database/sql"
+	"time"
+)
+
+type NullTime sql.NullTime
+
+func (nullTime *NullTime) ToTimeReference() *time.Time {
+	if nullTime.Valid {
+		return &nullTime.Time
+	}
+	return nil
+}
